move deprecation severity to core deprecation common package
