+++ src/core/packages/deprecations/common/tsconfig.json
@@ -12,8 +12,5 @@
   ],
   "exclude": [
     "target/**/*",
-  ],
-  "kbn_references": [
-    "@kbn/core-http-common",
   ]
 }
