+++ src/core/packages/deprecations/server-internal/src/deprecations_factory.test.ts
@@ -11,7 +11,7 @@ import type { GetDeprecationsContext } from '@kbn/core-deprecations-server';
 import { DeprecationsFactory, DeprecationsFactoryConfig } from './deprecations_factory';
 import { loggerMock } from '@kbn/logging-mocks';
 import type { DeprecationsDetails } from '@kbn/core-deprecations-common';
-import { DeprecationSeverity } from '@kbn/core-http-common';
+import { DeprecationSeverity } from '@kbn/core-deprecations-common';
 
 describe('DeprecationsFactory', () => {
   let logger: ReturnType<typeof loggerMock.create>;
