+++ src/core/packages/deprecations/server-internal/tsconfig.json
@@ -41,7 +41,7 @@
     "@kbn/core-logging-server-internal",
     "@kbn/core-doc-links-server",
     "@kbn/core-doc-links-server-mocks",
-    "@kbn/core-http-common",
+    "@kbn/core-deprecations-common",
   ],
   "exclude": [
     "target/**/*",
