+++ src/core/packages/http/common/index.ts
@@ -16,5 +16,4 @@ export {
   ELASTIC_INTERNAL_ORIGIN_QUERY_PARAM,
   X_ELASTIC_INTERNAL_ORIGIN_REQUEST,
   KIBANA_BUILD_NR_HEADER,
-  DeprecationSeverity,
 } from './src/constants';
