+++ src/core/packages/http/common/src/constants.ts
@@ -12,12 +12,6 @@ export const ELASTIC_HTTP_VERSION_HEADER = 'elastic-api-version' as const;
 export const ELASTIC_HTTP_VERSION_QUERY_PARAM = 'apiVersion' as const;
 export const ELASTIC_INTERNAL_ORIGIN_QUERY_PARAM = 'elasticInternalOrigin' as const;
 export const X_ELASTIC_INTERNAL_ORIGIN_REQUEST = 'x-elastic-internal-origin' as const;
-export enum DeprecationSeverity {
-  NONE = 'none',
-  INFO = 'info',
-  WARNING = 'warning',
-  CRITICAL = 'critical',
-}
 
 /** @internal */
 export const KIBANA_BUILD_NR_HEADER = 'kbn-build-number' as const;
