+++ src/core/packages/saved-objects/server-internal/tsconfig.json
@@ -47,7 +47,7 @@
     "@kbn/core-http-router-server-internal",
     "@kbn/logging-mocks",
     "@kbn/std",
-    "@kbn/core-http-common",
+    "@kbn/core-deprecations-common",
   ],
   "exclude": [
     "target/**/*",
