+++ src/core/packages/ui-settings/server-internal/tsconfig.json
@@ -33,7 +33,7 @@
     "@kbn/core-saved-objects-api-server-internal",
     "@kbn/core-saved-objects-common",
     "@kbn/utility-types",
-    "@kbn/core-http-common",
+    "@kbn/core-deprecations-common",
   ],
   "exclude": [
     "target/**/*",
