+++ src/platform/packages/shared/kbn-config/tsconfig.json
@@ -20,7 +20,7 @@
     "@kbn/i18n",
     "@kbn/doc-links",
     "@kbn/repo-packages",
-    "@kbn/core-http-common"
+    "@kbn/core-http-deprecations",
   ],
   "exclude": [
     "target/**/*",
