+++ src/platform/packages/shared/kbn-router-to-openapispec/tsconfig.json
@@ -20,6 +20,6 @@
     "@kbn/zod",
     "@kbn/zod-helpers",
     "@kbn/utility-types",
-    "@kbn/core-http-common",
+    "@kbn/core-deprecations-common",
   ]
 }
