+++ x-pack/platform/plugins/private/upgrade_assistant/tsconfig.json
@@ -44,7 +44,7 @@
     "@kbn/core-http-server-utils",
     "@kbn/core-elasticsearch-server",
     "@kbn/upgrade-assistant",
-    "@kbn/core-http-common",
+    "@kbn/core-deprecations-common",
   ],
   "exclude": [
     "target/**/*",
